refactor deposit withdraw page
